+++ CHANGES.rst
@@ -4,3 +4,4 @@ Changelog
 0.2
 ----
 - switched to using z3c.autoinclude to ensure security gets loaded after others
+- add utils helper to monkey patch methods that should only be used outside of restricted python